pyproject.toml
[project]
name = "mingpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.4",
    "python-dotenv>=1.2.1",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
